fix(ChatPrompt): extract IME composition guard into reusable composable
Move IME-aware Enter handling into `useIMEGuard` composable so users
building custom submit-on-enter inputs can reuse the Safari workaround.
Add behavioral tests for composition guard.
