+++ src/runtime/composables/index.ts
@@ -9,6 +9,7 @@ export * from './useFieldGroup'
 export * from './useFileUpload'
 export * from './useFilter'
 export * from './useFormField'
+export * from './useIMEGuard'
 export * from './useKbd'
 export * from './useLocale'
 export * from './useOverlay'
